pyproject.toml
[project]
name = "player"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.3.0",
    "pygame>=2.6.1",
    "pyqt6>=6.9.1",
]
